requirements.txt
emoji
aiohttp
aiofiles
ffmpeg-python
gTTS
dnspython
future
googletrans==3.1.0a0
motor
pymongo[srv]
pillow
psutil
pykeyboard
pyrogram
python-arq
requests
search_engine_parser
speedtest-cli
wget
TgCrypto
uvloop
bs4
pyyaml==6.0
python-dotenv
feedparser
fuzzysearch
pytube
validators
heroku3
apscheduler
#forsession
cachetools
regex
pytz
google_trans_new
dateparser
gpytranslate==1.4.0
EmojiCaptcha
captcha
rich
gitpython